+++ src/es6/core/frame.js
@@ -3,12 +3,12 @@
  * DataFrame object
  */
 
-// $FlowIssue
 import Immutable from 'immutable';
 // import { saveAs } from 'file-saver'; TODO figure out if best way
 
 import { InvalidAxisError } from './exceptions';
 import NDFrame from './generic';
+import { MultiIndex } from './multiindex';
 import Series from './series';
 // import { Workbook, Sheet } from './structs'; TODO
 import { enumerate, nonMergeColumns, intersectingColumns, parseIndex,
@@ -19,6 +19,7 @@ declare type T_LIST = Immutable.List
 declare type T_MAP = Immutable.Map;
 declare type T_SK = string|number;
 declare type T_COTHER = Array<T_SK>|T_LIST|Series|DataFrame|T_SK;
+declare type T_PVINDEX = Array<T_SK>|T_LIST|T_SK;
 
 
 const parseArrayToSeriesMap = (array: Array<Object>, index: T_LIST): T_MAP => {
@@ -265,8 +266,7 @@ export default class DataFrame extends NDFrame {
       const prevColumn = this.columns.get(idx);
       const prevSeries = this.get(prevColumn);
 
-      prevSeries.name = k;
-      nextData[k] = prevSeries;
+      nextData[k] = prevSeries.rename(k);
     });
 
     this._data = Immutable.Map(nextData);
@@ -327,7 +327,7 @@ export default class DataFrame extends NDFrame {
    * df.length;
    */
   get length(): number {
-    return Math.max(...this._data.keySeq().map(k => this.get(k).length).toArray());
+    return Math.max(0, ...this.columns.map(k => this.get(k).length).toArray());
   }
 
   /**
@@ -535,6 +535,11 @@ export default class DataFrame extends NDFrame {
    */
   tail(n: number = 10): DataFrame {
     return this.iloc([this.length - n, this.length]);
+  }
+
+  _assertColumnExists(col: T_SK) {
+    if (!this.columnExists(col))
+      throw new Error(`Column ${col} not in DataFrame`);
   }
 
   columnExists(col: T_SK): boolean {
@@ -1370,6 +1375,82 @@ export default class DataFrame extends NDFrame {
     return new DataFrame(data, {index: sortedIndex});
   }
 
+  /**
+   * Reshape data (produce a 'pivot' table) based on a set of index, columns, or values
+   * columns from the original DataFrame
+   *
+   * @param {Array<string>|Immutable.List|string|number} index
+   *  Name(s) of column(s) to use as the index for the pivoted DataFrame
+   * @param {Array<string>|Immutable.List|string|number} columns
+   *  Name(s) of column(s) to use as the columns for the pivoted DataFrame
+   * @param {Array<string>|Immutable.List|string|number} values
+   *  Name(s) of column(s) to use as the values for the pivoted DataFrame
+   * @param {string} aggfunc
+   *  Name of aggregation function
+   */
+  pivot_table(index: T_PVINDEX, columns: T_PVINDEX, values: T_PVINDEX,
+    aggfunc: string = 'sum'): any {
+    throw new Error('Not implemented');
+    const validateCols = (cols: T_PVINDEX): Immutable.List => {
+      if (Array.isArray(cols)) {
+        cols.forEach(c => this._assertColumnExists(c));
+        return Immutable.List(cols);
+      } else if (cols instanceof Immutable.List) {
+        cols.forEach(c => this._assertColumnExists(c));
+        return cols;
+      } else if (typeof cols === 'string') {
+        this._assertColumnExists(cols);
+        return Immutable.List.of(cols);
+      }
+
+      throw new TypeError('cols must be Array, Immutable.List, or string');
+    };
+
+    // Validate types and cast to Immutable.List of column names
+    const indexCols = validateCols(index);
+    const columnCols = validateCols(columns);
+    const valuesCols = validateCols(values);
+
+    let pivotMap = Immutable.Map({});
+
+    this.index.map((indexVal, idx) => {
+      const key = indexCols.map(c => this.get(c).iloc(idx))
+        .concat(columnCols.map(c => this.get(c).iloc(idx)));
+      let val = this.get(valuesCols.get(0)).iloc(idx);
+      if (pivotMap.has(key)) {
+        switch (aggfunc) {
+          case 'sum':
+            val += pivotMap.get(key);
+            break;
+          default:
+            throw new Error('not implemented for aggs');
+        }
+      }
+
+      // This pivotMap has indexCols.size keys then columnCols.size keys which point to the value
+      pivotMap = pivotMap.set(key, val);
+    });
+
+    let indexMap = Immutable.OrderedMap({});
+    let columnsMap = Immutable.OrderedMap({});
+
+    pivotMap.entrySeq().forEach(([k, v]) => {
+      const indexKey = k.slice(0, indexCols.size - 1);
+      console.log(k);
+      console.log(indexKey);
+      if (indexMap.hasIn(indexKey))
+        indexMap = indexMap.setIn(
+          indexKey, indexMap.getIn(indexKey).concat([k[indexCols.size - 1]]));
+      else indexMap = indexMap.setIn(
+        indexKey, Immutable.List.of(k[indexCols.size - 1]));
+      columnsMap = columnsMap.setIn(k.slice(indexCols.size, k.length));
+    });
+
+    console.log(indexMap);
+    console.log(columnsMap);
+    return pivotMap;
+  }
+
   _cumulativeHelper(operation: string = OP_CUMSUM, axis: number = 0): DataFrame {
     if (axis === 0) {
       return new DataFrame(
@@ -1460,6 +1541,23 @@ export default class DataFrame extends NDFrame {
    */
   cummin(axis: number = 0): DataFrame {
     return this._cumulativeHelper(OP_CUMMIN, axis);
+  }
+
+  /**
+   * Rename the `DataFrame` and return a new DataFrame
+   *
+   * pandas equivalent: [DataFrame.rename](http://pandas.pydata.org/pandas-docs/stable/generated/pandas.DataFrame.rename.html)
+   *
+   * @param {Immutable.Map} columns
+   * @returns {DataFrame}
+   */
+  rename({ columns }: {columns: Immutable.Map}): DataFrame {
+    return new DataFrame(Immutable.OrderedMap(this.columns.map((prevCol) => {
+      const nextCol = columns.get(prevCol);
+      if (typeof nextCol === 'undefined')
+        return [prevCol, this._data.get(prevCol)];
+      return [nextCol, this._data.get(prevCol).rename(nextCol)];
+    })), {index: this.index});
   }
 }